发布问题 用markdown插件 实现富文本编辑

+++ src/main/java/cn/duan/community/model/Comment.java
@@ -33,10 +33,7 @@ public class Comment {
 
     private Long gmtModified;
 
-
-    private Long likeCount;
-
-
+    private Integer likeCount;
 
     private Integer commentCount;
 }
